src/test/java/ru/job4j/threads/synch/SingleLockListTest.java
package ru.job4j.threads.synch;

import org.junit.Test;
import ru.job4j.synch.SingleLockList;

import java.util.ArrayList;
import java.util.Set;
import java.util.TreeSet;

import static org.hamcrest.CoreMatchers.is;
import static org.junit.Assert.assertThat;

/**
 * SingleLockListTest.
 *
 * @author fourbarman (maks.java@yandex.ru).
 * @version %I%, %G%.
 * @since 13.01.2022.
 */
public class SingleLockListTest {
    /**
     * Test when collection is ArrayList
     * and two threads add element to list
     * than list has 2 elements.
     */
    @Test
    public void add() throws InterruptedException {
        SingleLockList<Integer> list = new SingleLockList<>(new ArrayList<>());
        Thread first = new Thread(() -> list.add(1));
        Thread second = new Thread(() -> list.add(2));
        first.start();
        second.start();
        first.join();
        second.join();
        Set<Integer> rsl = new TreeSet<>();
        list.iterator().forEachRemaining(rsl::add);
        assertThat(rsl, is(Set.of(1, 2)));
    }
}
